Sorting default options in marker when building options_to_url_params
This is kinda stupid, but as the hashes are not sorted, and the test did
have a prefix order of something which it were expecting it kinda broke
in some situations, but in other it did work fine. So this will freeze
the order the options_to_url_params generate the string which should
mean that the test don't break again.