Map does no longer have @markers, in has @map_enteties.
@map_enteties will hold both markers and paths, and we use getter method
like markers and paths to filter out elements of the type which we want.

I did it like this, and not with two internal arrays, because I want to
be able to ask the map for it's first element, regardless of type. I
think this feelt more natural, than to select either marker or path to
be the first element returned by first.